refactor cli tools as a subpackage
